feat(agent): implement fetchAgentJobStatus method in AgentService and update job status retrieval in job service
